# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Movies API",
      "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
      "/register": {
        "post": {
          "tags": ["User"],
          "summary": "Register a new user",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User created"
            }
          }
        }
      },
      "/login": {
        "post": {
          "tags": ["User"],
          "summary": "Login a user",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Login"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User logged in"
            }
          }
        }
      },
      "/users": {
        "get": {
          "tags": ["User"],
          "summary": "Get all users with pagination",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "List of users"
            }
          }
        }
      },
      "/movies": {
        "get": {
          "tags": ["Movie"],
          "summary": "Get all movies with pagination",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "List of movies"
            }
          }
        },
        "post": {
          "tags": ["Movie"],
          "summary": "Add a new movie",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Movie"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Movie created"
            }
          }
        }
      },
      "/movies/{id}": {
        "put": {
          "tags": ["Movie"],
          "summary": "Update a movie",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Movie"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Movie updated"
            }
          }
        },
        "delete": {
          "tags": ["Movie"],
          "summary": "Delete a movie",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "204": {
              "description": "Movie deleted"
            }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Movie": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "genres": {
            "type": "string"
          },
          "year": {
            "type": "string"
          }
        }
      }
    }
  }
  